requirements.txt
anyio==3.6.2
APScheduler==3.9.1
async-generator==1.10
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
decorator==5.1.1
exceptiongroup==1.0.0
ffmpeg-python==0.2.0
future==0.18.2
greenlet==1.1.3
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
instaloader==4.9.5
outcome==1.2.0
pyee==8.1.0
PySocks==1.7.1
python-dotenv==0.21.0
python-i18n==0.3.9
python-telegram-bot==20.0a4
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tornado==6.2
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
wcwidth==0.2.5
wsproto==1.2.0
youtube-dl==2021.12.17
